requirements.txt
# Core Dependencies
Flask==2.3.2
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
flask_migrate>=4.0.0
python-dotenv==1.0.0

# Database Support (choose one based on your DB)
# PostgreSQL
#psycopg2-binary==2.9.6

# Security
#bcrypt==4.0.1
PyJWT==2.7.0

# API Support (if needed)
Flask-RESTful==0.3.9
Flask-Cors==3.0.10

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Linting & Formatting
flake8==6.0.0
black==23.7.0

# Development Extras
Flask-DebugToolbar==0.13.1
Flask-Testing==0.8.1
